Explore a little more structure in the client code.
Tried to improve the encapsulation of the child apps (i.e. pages and
menu component):

* Removed knowledge of App Msg DU from child apps, by putting the
App level message DU in a file after the child apps in the fsproj
compilation order.  Made it the responsibility of the App to map child
message types to App message types (done with composing the dispatch
function and the child msg ctor).  Done this under the assumption that
the child app should only know and communicate with its own messages.
* Moved the child app message types into the child app files.  Think
this is an easier read for newbies and is similar to lot of Elm
examples.

Other changes:

* Put Menu in a components folder, to reduce confusion about it being a
Page.  Seen this structure in other Elm apps.
* Put the 'global' model types in root Model.fs.  Not sure this is
quite right yet, for example, should a child component know about other
pages?  Anyway, OK for now.  Left the App specific model types in
App.fs.  This structure makesense in relation to the F# compilation order.
* Commented the page files as per Kris Jenkins suggestion of Model,
Message, REST, State, View.
See http://blog.jenkster.com/2016/04/how-i-structure-elm-apps.html