Try to work around Selenium problem by pinning some dependencies
UITests project was failing with the following message: `Only local connections are allowed. Error: Input string was not in a correct format.` (also saying `Connection refused` somewhere in there), indicating some kind of problem with newer versions of Selenium / the Chrome driver. So just work around it by forcing a pin to known working versions